PYCOLIMP.MNT: rename multilayers module.
