+++ pycolleff/pycolleff/impedances/__init__.py
@@ -4,12 +4,12 @@ from .coherent_sync_rad import CSRElement
 from .conv_wake_impedance import from_impedance_to_wake, from_wake_to_impedance
 from .element_and_budget import Element, Budget, load_budget, load_element
 from .kickers import kicker_coupled_flux, kicker_tsutsui_model
-from .multilayers import yokoya_factors, resistive_multilayer_flat_pipe, \
-    resistive_multilayer_round_pipe, RW_default_w, prepare_input_epr_mur, \
-    resistive_multilayer_round_pipe_multiprecision
+from .reswall_multilayers import yokoya_factors, prepare_input_epr_mur, \
+    resistive_multilayer_flat_pipe, resistive_multilayer_round_pipe, \
+    resistive_multilayer_round_pipe_multiprecision, RW_default_w
 from .resonators import longitudinal_resonator, transverse_resonator, \
     wake_longitudinal_resonator, wake_transverse_resonator
 from .transitions import taper
 
 del coherent_sync_rad, conv_wake_impedance, element_and_budget, kickers, \
-    multilayers, resonators, transitions
+    reswall_multilayers, resonators, transitions
